BUG: Use signals to capture delete events
